Add support for default values in attrs.
The default value is used to infer a type if no type is provided, and in the
generated __init__ function.

Note in particular that if a type *is* provided it is used unconditionally; the
type of the default is not checked against the type of the attr, so x =
attr.ib(type=str, default=10) will not raise an error.  (This is consistent
with current pytype behaviour for variables and function params with both
assignments and and type annotations.)

#pyfixit

PiperOrigin-RevId: 252167328