Remove pytype workarounds for previously missing/unparseable stubs.
This has been bothering me for a while. Typeshed has added quite a few new
stubs over the years, and pytype's parsing has also improved, but we've kept
around now-unnecessary stubs and entries in missing.txt. This cleans all that
up, fixes the tests, and resolves some lint warnings in the tests.

PiperOrigin-RevId: 261007907