Add an experimental --precise-return mode to pytype.
This mode causes pytype to use the first return type it finds in the case of a
failed function call, rather than Any. The immediate purpose is to improve
opcode tracing, so that type errors don't immediately cause type information to
be lost, but it's also a potential pytype improvement worth experimenting with.

PiperOrigin-RevId: 265510179