Look ahead up to 5 lines when tracing function calls.
The line number of the CALL_* opcode is the last line of the call (or the
second to last, if the last contains only a closing parenthesis), but the line
number of the ast node is the first line. The easiest workaround is to look
ahead a few lines for the opcode. I picked 5 because attribute lookahead uses
this value and it seemed as good as any.

#pyfixit

PiperOrigin-RevId: 269687977