Copy the numpy stub to pytype so it can be used externally.
As with the attr stubs, we have to keep two copies, one in pytype for
opensource users and one under third_party/py/ in order for that import path to
resolve properly.

Fixes https://github.com/google/pytype/issues/391 (well, the only part of it
that we can do anything about for now).

PiperOrigin-RevId: 271654867