Stop pinning clif_to_pyi to Python 3.6.
As part of the 3.6->3.7 upgrade work, I'm fixing places where pytype and
related tools assume 3.6 as the Python 3 version. clif_to_pyi is an easy case
because it does not use the minor version at all. We can just set the
--python_version flag to take a single int of the major version (2 or 3)
instead of a major.minor tuple.

The only tricky bit is that clif_to_pyi_test calls the pytype pyi parser, which
requires a (major, minor) Python version. Since I ultimately want to give
pytype the ability to take just a major version as well (so that we can say
--python_version=3 and have pytype use whatever the latest runtime version is),
I had clif_to_pyi_test pass the major version to pytype and added a new utils
method, full_version_from_major, to convert to a (major, minor) version. For
now, this method requires host Python 3 to work completely - optimistically,
I'll have pytype running under Python 3 by default by the time the rest of the
upgrade work is done. If not, I'll hardcode in a Python 3 version that this
method should return in Python 2.

I also removed some old and unnecessary logic in the parser to set the python
version to a default of (2, 7, 6) if a version of None was passed in.

PiperOrigin-RevId: 279379968