Rename pytype.main as pytype.single
Having pytype.main actually running pytype-single is somewhat confusing.
This situation is a remains from the time the pytype executable was
what now is pytype-single (see e.g. commit 3de77436). Having a
"pytype.single" module clarifies the situation.

The underlying motivation is to make it possible to run pytype-single as
`python -m pytype.single` (following the move of "pytype" executable in
commit 79f8d35d93f8798d3bf7be0120c8a058a8af5280).