Simplify vm.local_traces and vm.ordered_locals.
Previously, vm.local_traces was used to get the attribute order for dataclasses
but also recorded attribute values that weren't used, and vm.ordered_locals was
kept in the right order for attrs.

Instead, vm.local_traces has been changed to vm.local_ops, which records
variable names and operations (ANNOTATE or ASSIGN) in the order in which
they're done, and vm.ordered_locals has been changed to an unordered dict of
variable names to values. classgen.Decorator.get_class_locals takes a new
argument, ordering, which can be used to request either the attrs or the
dataclasses ordering.

I also fixed a bug in the behavior of attr.s(auto_attribs=True): only annotated
attributes should be picked up and should follow the dataclasses ordering.

PiperOrigin-RevId: 282402682