Optimize bytecode compilation when the host and target Python versions match.
Currently, when we want to compile source code, we write it to a temporary
file and spawn a subprocess to call compile_bytecode, which reads the file.
When the host and target Python versions match, we can call compile_bytecode
directly, bypassing both the file write+read and the subprocess call.

#pyfixit

PiperOrigin-RevId: 283675427