Overhaul the LateAnnotation implementation.
Significantly changes how forward references are implemented, reducing the
number of places where they need to be special-cased and fixing a number of
bugs and edge cases.

Unfortunately, to avoid breaking our existing forward reference support, I had
to make this change all at once - apologies for the mega-CL.

The way the new late annotations work is that when a forward reference is
encountered, we immediately evaluate it but set a special flag on the VM that
tells it to generate LateAnnotation objects instead of reporting name errors.
Then, whenever a new global is defined, we resolve any late annotations that
refer to it. Once all globals have been defined, we resolve any remaining
annotations so that real name errors are still reported.

The LateAnnotation class defines __getattribute__, allowing it to pretend to be
an unsolvable before resolution and its resolved type afterwards.

I expect this change to break a bunch of targets that have type errors in their
forward references or function type comments, so I won't submit it until we've
sorted out the typeshed update.

Fixes https://github.com/google/pytype/issues/420.

Startblock:
  b/143674618 is fixed
PiperOrigin-RevId: 284678790