Further modernize python build system (#509)
* Replace conditional logic with environment markers

PEP 496 introduced environment markers, which allow you to declare a
dependency conditional on the environment the package is to be installed
into. Among other things, this is useful for cross-compilation.

* Add Python 3.7 to trove classifiers

* Add python_requires matching test matrix

This will prevent `pip` from finding versions of pytype that do not
support a given version of Python. Unfortunately, when no
python_requires declaration is made, the default is "acceptable for all
versions", and so without yanking old versions, `pip install pytype` on
unsupported versions will simply fall back to the latest release without
a `python_requires` declaration. Alas, this still seems like a win.

* Add Python 3.5 to trove classifiers

* Move all static build options into setup.cfg

Using the declarative build file for all your declarative options is
current best practice, since it makes it much easier to parse metadata
from projects without executing arbitrary code.

This will change the minimum required version of setuptools to 30.4,
which was released in December 2016.

* Add pyproject.toml file with build-system

This declares the build system and (Python) dependencies required to
build the project, and will trigger PEP 517 builds when present in pip
installs.

* Add pyproject.toml to MANIFEST.in

This will include pyproject.toml in sdists, which will trigger PEP 517
builds when installing from sdist.

* Add license to build metadata

* Use find_packages to populate the package list

This will prevent further issues like #372 (though technically #372
would still have caused a problem because __init__.py was missing).

* Add project_urls

This allows PyPI to give multiple links to various aspects of the
project. See an example of how this renders on urllib3's PyPI page:
https://pypi.org/project/urllib3/1.25.8/

* Move version into setup.cfg

Importing from a fixed module is a built-in capability of the
declarative build file, so no need to re-invent it ourselves.