Use //devtools/python/launcher:python3_binary directly in pytype.
This change allows us to get rid of one of our copies of the Python
interpreter. It will also make version upgrades easier because pytype will
follow the runtime Python 3 version rather than being pinned to whatever we've
most recently copied.

We can't do the same thing with :python2_binary because the type annotations
backport requires the binary to be built with -c opt.

Startblock:
  b/147668882 is fixed
PiperOrigin-RevId: 302102449