Remove dependency on now-unnecessary python3_binary.
Now that pytype no longer has host Python 2 tests, we should not need a custom
Python 3 binary.

I noticed that pytype_source_utils_test was missing from CMakeLists.txt, so I
added it.

If this change works, that'll be confirmation that we managed to migrate fully
to Python 3 =)

For https://github.com/google/pytype/issues/545.

PiperOrigin-RevId: 304831817