Test against 3.8 and master with allowed failures.
This adds non-required jobs that will test against 3.8 and the master
branch, which should make it easier to determine what needs to be done
to add support for those Python versions.

Note: This will currently always fail trivially, because it's excluded
in python_requires. This will still be beneficial because
non-maintainers can modify setup.cfg in their PR branches, but they
cannot modify .travis.yml. A better long-term solution would be to add a
workaround that allows installing the package on nominally unsupported
versions of Python.