Make (func: T) -> T work properly in the same file.
Previously, when a function was annotated like this, the annotations worked
properly when the function was used in a different file but not the same file.

I changed the file's linting to py3 so I could use super() with no arguments
and fixed a couple of new warnings about deprecated inspect methods.

PiperOrigin-RevId: 305828065