Fix: handle typing.Type in convert.py.
When a module re-exports a type from another one:
  X = other_mod.X
what ends up being written to the pyi file is:
  X: Type[other_mod.X]

pytype was crashing on this. I couldn't reproduce the problem locally, so it
might have something to do with pickling and late types. I confirmed that the
crash disappears after this fix.

PiperOrigin-RevId: 311795396