Replaces values with Any when a variable is annotated as Any.
In general, we can't use annotations to replace values because this messes up
things like if-splitting. But specifically in the case that a variable is
annotated as Any, we've gotten bug reports that suggest that it is surprising
to users when pytype builds a Union[Any, ...] based on further assignments to
the variable.

PiperOrigin-RevId: 313439881