Partial rollback of changelist 318522426.
*** Reason for rollback ***

The PyPI release is out, so revert temporary changes made for the release.

*** Original change description ***

Do a PyPI release for a parser fix.

For https://github.com/google/pytype/issues/608.

I fixed a pyi parser issue about two weeks ago, and we need to do a release so
that typeshed can use the fix.

I don't feel confident releasing at head (too many findings to click through to
check if there are any new crashes, obvious bugs, etc.), so what I plan to do
is to temporarily roll back the change that generates the vast majority of the
new findings, release on PyPI, and then reapply that change.

***

PiperOrigin-RevId: 318553012