Make the behavior of io.open match builtins.open.
io.open is an alias for builtins.open, but pytype was treating them differently
because builtins.open was implemented using an overlay. In turn, the overlay
was needed because pytype would incorrectly match constants to both Literal and
non-Literal annotations. This change improves pytype's Literal matching,
allowing us to remove the overlay, which fixes the buggy behavior of io.open.

PiperOrigin-RevId: 319260037