Fix a weird bug involving calling __getitem__ on unions.
The CL to support Union type macros caused a bizarre bug where
Union.set_slot('__getitem__') calls getattribute(Union, '__getitem__'),
creating an extra node, causing convert_class_annotations for a tuple
annotation containing a forward referenced imported type to be evaluated at the
wrong node, leading to the generation of an extra late annotation for the
imported module, finally producing "Invalid type annotation '<instance of
module>': Not a type" errors all over the place for pb2 enum annotations, since
those always have to be quoted. I resolved this by special-casing
Union.__getitem__ in attribute.py, so the extra node is not generated.

I also noticed that the type macro CL produced a corner case where
Union.getitem_slot is indistinguishable from calling __getitem__ on a Union's
options when a value is annotated as Type[Union[C1, C2, ...]] (but *not* when
the annotation is Union[Type[C1], Type[C2], ...]!). The inferred type is still
correct, if imprecise, so I just documented this behavior in a test rather than
trying to fix it.

PiperOrigin-RevId: 324684760