Small refactoring to parser.py so we can patch it to read files from CNS.
We want to run pytype, while generating the PYI files tree in CNS. We do it because we work on known historical [] snapshots, so once we generated a PYI file for a snapshot we can keep it forever (or at least until we change our code).

The built in python "open" cannot read from CNS. So we patch the read function to use gfile.open which is a drop-in replacement.
Like we do here:
[]

PiperOrigin-RevId: 328780052