Distinguish more carefully between command-line and library options.
The config.Options object is used to parse and store both command-line and
library options. This hasn't mattered up to now because these two sets of
options were the same. But I'd like to add a library-only option (the ability
to replace builtins.open with a custom open function). With this change,
command-line utilities will be expected to construct Options objects as
Options(argv, command_line=True) and libraries will be expected to use
Options.create(input_filename, **options).

PiperOrigin-RevId: 328996421