Pass around an options.open_function that is used in place of builtins.open.
Adds a open_function option to config.Options, which can be set to drop-in
replacement for builtins.open that is used everywhere pytype reads or writes a
filepath provided by the user.

I added a couple spot tests that open_function is being used in various places
but did not test exhaustively.

PiperOrigin-RevId: 329000803