Fix a SETUP_FINALLY bug in Python 3.8.
In 3.8, SETUP_EXCEPT is removed and SETUP_FINALLY does the jobs of both
SETUP_EXCEPT and the old SETUP_FINALLY. We can figure out which to call based
on whether the opcode targets an except or finally block. Since the opcode
refers to blocks by id, OrderedCode now maintains an {id: block} dictionary
rather than a list of blocks.

For https://github.com/google/pytype/issues/440.

PiperOrigin-RevId: 329759870