Print 'leaving directory' after ninja command.
ninja prints 'entering directory', but the directory does not actually change,
which causes problems for emacs users. We want to keep the ninja output, since
it includes status updates for the build, so the easiest way to fix this is to
print 'leaving directory' ourselves.

Fixes https://github.com/google/pytype/issues/659.

PiperOrigin-RevId: 330629292