Add more Python 3.8 support.
With this change, all of the crashes in pytype's 3.8 tests are gone; there are
only a handful of normal failures left. I've gone in circles a bit (you may
notice that I reverted parts of a couple of earlier changes), but given that
the tests are in a much better state now, I'm hopeful that I've finally got
things more or less correct.

The one oddity is that SETUP_FINALLY appears to need to push the exception onto
the value stack twice. A consistent issue I was seeing was the VM crashing due
to ROT_FOUR only having one value to rotate. I got the idea to push an extra
three-value exception from seeing that the CPython code does it
(https://github.com/python/cpython/blo[]b01b58ee1af821611cc436adb20f3e8b879d/Python/ceval.c#L3736),
but that appears to happen in 3.7 as well, so I could be totally off-base.

For https://github.com/google/pytype/issues/440.

PiperOrigin-RevId: 331009065