Fix a parser bug when importing and re-exporting a submodule.
load_pytd distinguishes between the names defined in a packages's __init__ file
and submodules based on whether the submodule name is present in __init__. This
breaks when a submodule is re-exported under the same name because __init__
contains Alias('module.submodule', NamedType('module.submodule')). The easiest
fix is to not create the alias in the first place, by treating a re-export
under the same name as a simple import.

While debugging, I noticed a bug where we handled renamed modules incorrectly,
so I fixed that as well.

PiperOrigin-RevId: 331637694