Port function signature code from parser.py to the typed_ast parser.
Also fixes a bug with updating array nodes in traces/visitor.py and fixes a
function in xref that was relying on the buggy behaviour.

Right now I'm testing by comparing the output of the two parsers; I'll add unit
tests once the typed ast parser is more complete and we have a finalised design.

PiperOrigin-RevId: 332093271