Switch the Travis build environment to Focal Fossa.
My previous cmake change worked fine locally but failed on Travis once I
tried to pull in a no-srcs cc_library. Turns out that the cmake version
in most of Travis's build environments is quite old, but Focal Fossa is
nice and modern (so modern that we don't need to separately install
Python 3.6 and 3.7 but do need to install 2.7!).