Use `py -x.y` instead of `pythonx.y` for windows.
Although pytype doesn't support windows, the typeshed pytype_test used
to work under windows when --python_exe could be used to pass in the
path to the python interpreter. I broke the test when I removed this
flag (oops). Attempting to fix it by doing a sys.platform check and
using the python launcher for windows
(https://docs.python.org/3/using/windows.html#python-launcher-for-windows).

I tested this out in a 64-bit Windows 10 VM. I wasn't able to get pytype-single
completely working in Windows (kept running into a parser error that I think
was caused by either the settings with which the C extension was built or
Windows doing some weird encoding...something to my test files), but it worked
well enough that I'm ~80% sure that pytype is able to find the Python
interpreter after this change.

Closes https://github.com/google/pytype/pull/682.

PiperOrigin-RevId: 334270850