Fix definition of ByteString and the `bytes` shorthand.
typing.ByteString should represent bytes, bytearray, and memoryview, with
bytes serving as a shorthand for all three types when it appears as a parameter
annotation. The simplest way to do this is to cheat slightly and represent
ByteString as a Union instead of a single type. To make the bytes shorthand
work properly, I also had to add some logic in the matcher for deduplicating
compatible builtins when matching typevars.

PiperOrigin-RevId: 336705554