Add pytype/typegraph/* to MANIFEST.in.
I was getting errors about cfg_logging.h being missing when trying to
test a new release; this change seems to fix the issue.

I'm still seeing some odd dynamic import failure, but python3.7 on my
machine also appears to be half-broken right now, so I have no idea if
the issue is with the pytype code or my environment.