Fix: don't rely on the module attribute in analyze._is_typing_member.
The module check broke some user code that was importing a typing member, then
exporting a parameterized version under the same name.

As a side note, I noticed that pytype infers a type of Any for TypingContainer
objects. I didn't try to fix this, since I can't think of a good reason to want
to re-export a typing member from your own module and use that instead of using
typing directly.

PiperOrigin-RevId: 343207525