Take classmethods and staticmethods into account in protocols.
An unexpected consequence of supporting typing_extensions.runtime is that we
suddenly analyze a bunch of code that previously silently passed type-checking
due to the use of a runtime protocol. This runtime protocol also has a
classmethod, which caused a bunch of false positives due to pytype
inconsistently ignoring classmethods some of the time.

PiperOrigin-RevId: 343342498