Cleanup: create a LazyMembers mixin.
Most of the logic for lazy attribute loading was defined in the
SimpleAbstractValue class, but SimpleAbstractValue is not lazy, so the class
contained a bunch of code irrelevant to its own functionality. I split out the
lazy loading code into a new LazyMembers mixin, which also allowed me to get
rid of the is_lazy attribute in favor of an isinstance check.

I also got rid of the usage of six in mixin.py and replaced another attribute,
CAN_BE_ABSTRACT, with an isinstance check as well.

PiperOrigin-RevId: 343427297