Treat classes whose metaclass inherits from type(Protocol) as abstract.
See linked bug for motivation. I first tried changing the definition of
Protocol in typing.pytd to include a new _ProtocolMeta metaclass, but this
broke a lot of tests because Protocol is a base class for so many typing
objects, and a custom metaclass screwed up attribute handling in some corner
cases. Special-casing Protocol when accessing type() and __class__ is a less
invasive change.

PiperOrigin-RevId: 345489232