Fix a [name-error] caused by a bad regex check.
We were expecting the name of a list comprehension to be '<listcomp>'. In
Python 3, the name has the scope prefixed, so the name of a list comprehension
inside a module-level function foo is 'foo.<locals>.<listcomp>'. Due to us not
recognizing the list comprehension, we analyze it as a stand-alone function,
which means that in the linked issue, it gets analyzed after `del num` has run,
leading to a [name-error].

Fixed by using re.search instead of re.match, so we check that the name ends
with '<listcomp>' rather than that the entire name is '<listcomp>'.

Fixes https://github.com/google/pytype/issues/750.

PiperOrigin-RevId: 347414681