Fix crash caused by trying to unpack a namedtuple.
This is what was causing the last "Closure lookup failed" error. At first
glance, it looks like a namedtuple bug - the instance type parameter ought to
be populated. But I think adding an Any binding to it is the approach least
likely to break anything while we're trying to get a release out the door.

PiperOrigin-RevId: 347736406