Support using a union to filter types out of a type parameter.
For a function signature like:
  def f(x: Optional[T]) -> T: ...
you'd expect that if you pass f a value that can be None, None will be filtered
out of the possible return values.

Instead, pytype was allowing None to match both None and T in the annotation.
Fixed by not allowing a value to match both a type parameter and a non-type
parameter.

Resolves https://github.com/google/pytype/issues/752.

PiperOrigin-RevId: 348543825