Support redefining NamedTuple in the old pyi parser.
This change is *almost* pointless, given that the new parser handles this just
fine and is about to be merged in, but it's a trivial fix and will unbreak a
Google opensource project that I just noticed no longer type-checks after the
removal of backticks. Plus the test is good to have.

PiperOrigin-RevId: 351872790