Export the typed_ast pyi parser to github.
This parser is intended as a drop-in replacement for the current yacc-based
parser. Rather than maintaining our own grammar, it uses the standard typed-ast
library to parse pyi files as python3 code, before converting them to pytype's
internal data structures. This fixes several issues with pytype not parsing
some constructs in typeshed, and should help us track future python syntax
changes a lot more easily.

Since this has several conflicts with the existing parser, it has been
developed internally in a separate branch, and is being merged in as a unit. We
will add detailed documentation of its internals to our developer docs.

PiperOrigin-RevId: 352670366