Support typeshed's new directory structure.
See https://github.com/google/pytype/pull/821 and its linked issue for context.
The typeshed directory structure is changing significantly, so we need to
update pytype accordingly.
https://github.com/python/typeshed/issues/2491#issuecomment-758772838 contains
a nice diagram of the new structure.

Note that I first developed this change on GitHub, then imported the PR. I'm
asking for a review on the import (rather than the PR) because the import
contains additional BUILD file changes (especially to third_party/py/toml - see
the diffbase).

PiperOrigin-RevId: 354138398