Partially support wrapping attr.s.
With this change, pytype will recognize wrappers for attr.s as long as they're
used in the same file in which they're defined.

I checked our other uses of Args.simplify() and realized that we typically
simplify the arguments before matching, not after as I did in my previous
dataclasses change, so I fixed that as well.

PiperOrigin-RevId: 355544594