Fix an if-splitting bug by (mostly) ignoring cycles in the CFG.
When an element is added to a container, followed by more elements added in a
loop, pytype would forget a None check done on the first element. The reason
for this is that, when the additional elements are added, the None option for
the first element is copied to one of the nodes in the loop, and then the
binding in the loop is considered always solvable because it's treated as its
own origin. This can be fixed by ignoring cycles except when there are no other
possible origins.

This fixes 6 (!!!) user-reported issues.

PiperOrigin-RevId: 357738294