Fix a bug in how pytype detects nested generic classes.
Previously, pytype was checking a function's locals and treating any
InterpreterClass found there as a definition of a nested class. This meant that
any reference to a generic InterpreterClass was incorrectly treated as a
definition whose template needed to be checked. Instead, we should only check a
local if it is a class that is stored under its own name.

PiperOrigin-RevId: 358223202