Don't report [not-instantiable] when instantiating an abstract annotation.
pytype helpfully reports [not-instantiable] if you try to create an instance of
an abstract class. However, this error is not helpful if the abstract class
comes from a `Type[AbstractClass]` annotation, as it's a common pattern to
annotate a function as accepting a `Type[AbstractClass]` and pass in
non-abstract subclasses. It suddenly occurred to me today that fixing this bug
is trivial now that we have a BaseValue.from_annotation attribute =)

PiperOrigin-RevId: 359580391