Move dataclass code generation to PyTDClass.__init__ to support inheritance.
* Moves pytd code generation directly into pytd/codegen/decorate, since we only
  want this for dataclass-like decorators right now, and attrs will use the
  same code.
* Adds support for one pyi dataclass inheriting from another
* Validates pyi dataclasses in the parser.

This code is pretty specific to the dataclass use case (generating an __init__
function from fields collected from all parent dataclasses), and can be
generalised if we ever want to support other pyi decorators that need inherited
attributes. (Code generation that does not need inherited attributes can
continue to be done in the parser, e.g. namedtuples).

Some of the code is pretty hacky due to the difficulty of generating "final"
pytd subtrees without going through the loader.

PiperOrigin-RevId: 362612365