Allow annotation-type-mismatch directives to be put on more sensible lines.
I've been cleaning up --check-parameter-types findings, to try to bring the
number of findings down to something manageable before release. An issue that
has come up several times is that it's confusing when the disable directive
appears in the middle of a function signature, since it misleads the reader
into thinking the directive applies only to the parameter(s) on that line.

This change allows the directive to be put on any of the lines between the line
of the MAKE_FUNCTION opcode (inclusive) and the line of the next opcode
(exclusive).

PiperOrigin-RevId: 363790579