Allow reassigning annotated parameter names.
--check-parameter-types contained two kinds of checks:
(1) checking that a parameter annotation matches its default.
(2) checking that when an annotated parameter is reassigned, the new value
    still matches the annotation.

(1) is frequently requested and something we definitely want to do. (2) is very
disruptive and has not actually been asked for by users, so I've removed it. We
may or may not want to add this check down the road, but I don't think it
should be part of --check-parameter-types, especially since (1) already causes
1000+ new findings.

I've kept the behavior that when a parameter is annotated as Any, its type
stays as Any rather than Union[Any, <new types assigned in the function body>],
since an explicit Any annotation often indicates that the user wants to
override inference.

PiperOrigin-RevId: 364709605