Fix typeshed handling for .par pytype installations.
When pytype is packaged as a .par, os.path.exists and os.listdir don't work,
breaking our handling of the new typeshed structure. After this change, I
believe everything is (finally!) working for both par and non-par
installations; I ran the tests, confirmed that the right number of typeshed
modules is loaded, and built some pytype targets both in dev mode and with
rebuilt pars.

* Adds _file_exists and _list_files methods that work for both par and non-par
  pytype installations.
* Reworks _load_third_party_packages to do a single recursive walk of all files
  in stubs/. For a par installation, we can't (efficiently) listdir directories
  to get only the names immediately in a directory.
* Tweaks get_all_module_names to remove the complexity of calling a
  _build_lister wrapper to build a lister function; instead, the internal
  lister function can now be used directly.
* In get_all_module_names, handle typeshed and pytype/stubs/ paths separately.
  This makes the file handling code simpler because it only has to deal with
  typeshed paths.

PiperOrigin-RevId: 368044952