Substitute in the right typevar values in children of generic classes.
When a non-generic class inherits from a generic one, the child does not have a
(non-empty) template but one of its parents does. The logic to substitute in
the right TypeVar values in method calls works by temporarily setting a `self`
annotation when a template is present, so we need to walk up the class's MRO
looking for the generic parent.

PiperOrigin-RevId: 368782058