Add a vm state to allow recursive construction of pytd values.
This is needed for the single use case of dataclass code generation in an
imported PyTDClass, but it is hard to handle recursive field annotations in any
other way.

Also removes the class-to-any visitor, which was an earlier attempt to handle
recursion, but which did not cover all the corner cases.

PiperOrigin-RevId: 372440646