Fix a Python 3.7 failure in typeshed/tests/pytype_test.py.
The failure is triggered when concurrent.futures.process loads
multiprocessing.context which loads multiprocessing, which isn't able to
resolve multiprocessing.context references yet. Instead, we directly look up
the references in the multiprocessing.context AST.

I unfortunately wasn't able to construct a local test case for this error
because it seems to depend on pytype_test loading files in a specific order,
but this code path will be exercised by the test once it is enabled in 3.7.

PiperOrigin-RevId: 372649664