Fix a generics bug.
Pytype was inferring a weird type for `self` in a generic class when (1) the
TypeVar has constraints, *and* (2) the class has a subclass defined in the same
file. Fixed by filtering out the call record that introduces the bad type. I
had to do this rather than not recording the call in the first place because
pytype sometimes does and sometimes doesn't want to record method calls from
subclasses - we can't tell until we've collected all of the call records.

Also fixed a small bug where the pytd printer was sometimes adding unnecessary
imports, which was triggered by the new test I added in this change.

PiperOrigin-RevId: 372981555