Add support for calling metaclasses from overlays.
This is needed for #pytype_enums. It may not be entirely necessary to pass in the name, because that should be on the class object that's passed in as the first argument. The full list of bases is absolutely needed, because e.g. enums process the list of bases and use it when constructing the enum. It's still not clear if the last argument is needed.

PiperOrigin-RevId: 373828024