Fix a performance regression in pytype/pytd/typeshed.py.
Fixes a performance regression caused by
https://github.com/google/pytype/pull/906 calling self._file_exists()
repeatedly, which is quite slow in .par files. This wasn't caught before
release because (1) the issue only appears when pytype is run from a .par file,
and (2) we usually load typeshed once, pickle it, and never load it again, so
the increase in the time needed to do the pickling was barely noticeable.
However, when pytype.par is run on its own, it takes more than 3x longer to
analyze a simple file due to this issue.

Fixes the test timeout in the attached bug.

PiperOrigin-RevId: 374928257