load_pytd: tweak when log_module_not_found is called.
log_module_not_found logs a warning when a module is not found, which causes
test failures when I try to import the latest version of typeshed because
single_test expects stderr to be clean.

I don't want to remove log_module_not_found altogether because I've sometimes
found its output useful in the past, but it shouldn't log anything in cases in
which not finding a module is fine and expected.

PiperOrigin-RevId: 377444785