Add a constant folding optimisation before running the bytecode.
This optimisation collapses the bytecode to construct a literal constant into a
single LOAD_FOLDED_CONST opcode, with the type of the literal attached to it.
There is corresponding code in the vm to convert the literal into a vm type and
put it on the stack.

This is a work in progress; the major missing feature is preserving the actual
python value of the constant.

PiperOrigin-RevId: 378752757