Support non-method attributes in typing.Protocol.
With this change, pytype detects and enforces non-method attributes in
protocols. There's one remaining piece of work, which is to add more details to
error messages to make it more obvious why something didn't match a protocol.

Somewhat surprisingly, this change appears to cause only one new finding, so it
doesn't need a feature flag.

PiperOrigin-RevId: 381117986