Allow TypeVars to be passed to functions.
Gets rid of the 'cannot pass a TypeVar to a function' error in favor of
replacing TypeVars with 'object' in the matcher. This lets us handle TypeVars
with minimal changes to the matching code.

This lets us delete some code that was put in to work around the TypeVar error.

Getting rid of annotations_util.deformalize() also exposed a bug where we
didn't allow annotations like Sequence[object] to be instantiated, which I
fixed by improving ParameterizedClass._is_callable().

For https://github.com/google/pytype/issues/717: This fixes the
"invalid-typevar" error reported in that issue but not the "not-callable" one.

PiperOrigin-RevId: 383791092