Improve [name-error] message and add skipped tests for future work.
When a local variable name shadows a global one, adds an explanation to the
[name-error] message explaining why the name cannot be referenced before being
locally defined.

Also adds some skipped tests for other cases in which we do not yet add this
explanation, since I haven't yet figured out how to look up a name in outer
class and function scopes.

PiperOrigin-RevId: 385624371