Support generic namedtuples.
* Adds the ability to generate mutations from __new__ methods, since NamedTuple
  uses __new__ rather than __init__.
* Passes the list of bases to NamedTuple.make_class() so that the Generic base
  is detected.
* Ensures the class's __annotations__ dict is initialized properly (somewhat
  tricky when the NamedTuple is constructed via a function call) - once that's
  done, our existing code for handling generic classes kicks in and takes care
  of the rest.

As a side note, using generic namedtuples seems rather inadvisable, since
whether they work at runtime changes from version to version. This fails in
Python 3.6:

>>> from typing import Generic, NamedTuple, TypeVar
>>> T = TypeVar('T')
>>> class Foo(NamedTuple, Generic[T]):
...   x: T
...
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: metaclass conflict: the metaclass of a derived class must be a (non-strict) subclass of the metaclasses of all its bases

succeeds in 3.7, and is broken again in 3.9:

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/lib/python3.9/typing.py", line 1881, in _namedtuple_mro_entries
    raise TypeError("Multiple inheritance with NamedTuple is not supported")
TypeError: Multiple inheritance with NamedTuple is not supported

But that's outside of our control, and if someone is really determined to
create a generic namedtuple, they can do it in a TYPE_CHECKING block.

PiperOrigin-RevId: 385710726