Get rid of a spurious 'appear only once in signature' TypeVar error.
Note that this fix depends on the 'closure' attribute of InterpreterFunction
holding an instance of 'self' for functions nested inside instance methods.
This should always be the case when the nested function's annotations reference
the class's type parameters, since you can't get the value of a type parameter
without accessing 'self'.

PiperOrigin-RevId: 386086421